requirements.txt
dedalus-labs
python-dotenv
fastapi
uvicorn

